requirements.txt
numpy
pygame
numpy-quaternion
xlrd
scipy
tensorflow
matplotlib
